+++ bls/Cargo.toml
@@ -9,6 +9,17 @@ homepage = { workspace = true }
 license = { workspace = true }
 edition = { workspace = true }
 
+[lib]
+crate-type = ["rlib"]
+
+[features]
+bytemuck = ["dep:bytemuck"]
+default = ["std"]
+frozen-abi = ["dep:solana-frozen-abi", "dep:solana-frozen-abi-macro", "std"]
+serde = ["dep:cfg_eval", "dep:serde", "dep:serde_with"]
+solana-signer-derive = ["dep:solana-signer", "dep:solana-signature", "dep:subtle"]
+std = ["dep:serde_json"]
+
 [dependencies]
 base64 = { workspace = true }
 bytemuck = { workspace = true, optional = true }
@@ -16,10 +27,10 @@ cfg_eval = { workspace = true, optional = true }
 serde = { workspace = true, features = ["derive"], optional = true }
 serde_json = { workspace = true, optional = true }
 serde_with = { workspace = true, features = ["macros"], optional = true }
-solana-frozen-abi = { version = "2.2.1", optional = true, features = [
+solana-frozen-abi = { workspace = true, optional = true, features = [
     "frozen-abi",
 ] }
-solana-frozen-abi-macro = { version = "2.2.1", optional = true, features = [
+solana-frozen-abi-macro = { workspace = true, optional = true, features = [
     "frozen-abi",
 ] }
 thiserror = { workspace = true }
@@ -30,8 +41,8 @@ blstrs = { workspace = true }
 ff = { workspace = true }
 group = { workspace = true }
 rand = { workspace = true }
-solana-signature = { version = "2.2.1", optional = true }
-solana-signer = { version = "2.2.1", optional = true }
+solana-signature = { workspace = true, optional = true }
+solana-signer = { workspace = true, optional = true }
 subtle = { workspace = true, optional = true }
 
 [dev-dependencies]
@@ -39,21 +50,6 @@ bincode = { workspace = true }
 solana-bls = { path = ".", features = ["std"] }
 solana-keypair = { workspace = true }
 tempfile = { workspace = true }
-
-[features]
-bytemuck = ["dep:bytemuck"]
-default = ["std"]
-frozen-abi = ["dep:solana-frozen-abi", "dep:solana-frozen-abi-macro", "std"]
-serde = ["dep:cfg_eval", "dep:serde", "dep:serde_with"]
-solana-signer-derive = [
-    "dep:solana-signer",
-    "dep:solana-signature",
-    "dep:subtle",
-]
-std = ["dep:serde_json"]
-
-[lib]
-crate-type = ["rlib"]
 
 [lints]
 workspace = true
